requirements.txt
black
flake8
pytest
mypy
docformatter
pre-commit
pydocstyle==6.1.1
deep_translator

accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.0.9
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
datasets==2.19.0
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.2
exceptiongroup==1.2.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pip==23.3.1
pluggy==1.5.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
PySocks==1.7.1
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.19.0
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.65.0
transformers==4.40.0
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.1.0
wandb==0.16.6
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
